メモリの持ち方を変更、メモリアドレス予約のバグ修正、CSRのクラス変更
