+++ RV32_Lsu/LoadStoreUnit/RV32_AmoLsu.cs
@@ -8,7 +8,7 @@ namespace RiscVCpu.LoadStoreUnit {
     /// <summary>
     /// Risc-V RV32A 拡張命令セット アトミック(不可分)なメモリ操作に対応したロードストアユニット
     /// </summary>
-    public class RV32_AmoLsu : RV32_Lsu {
+    public class RV32_AmoLsu : RV32_IntegerLsu {
 
 
         /// <summary>
@@ -31,8 +31,8 @@ namespace RiscVCpu.LoadStoreUnit {
         /// <param name="rs2">レジスタ番号</param>
         public bool LrW(Register rd, Register rs1, bool acquire, bool release, UInt32 insLength = 4u) {
             UInt32 addr = reg.GetValue(rs1);
-            if (mem.CanOperate(addr, 4)) {
-                mem.Acquire(addr);
+            if (reg.Mem.CanOperate(addr, 4)) {
+                reg.Mem.Acquire(addr, 4);
                 base.Lw(rd, rs1, 0);
             } else {
                 reg.IncrementPc(insLength);
@@ -49,9 +49,9 @@ namespace RiscVCpu.LoadStoreUnit {
         /// <param name="rs2">レジスタ番号</param>
         public bool ScW(Register rd, Register rs1, Register rs2, bool acquire, bool release, UInt32 insLength = 4u) {
             UInt32 addr = reg.GetValue(rs1);
-            if (mem.CanOperate(addr, 4)) {
+            if (!reg.Mem.CanOperate(addr, 4)) {
                 base.Sw(rs1, rs2, 0);
-                mem.Release(addr);
+                reg.Mem.Release(addr);
                 reg.SetValue(rd, 0);
             } else {
                 reg.IncrementPc(insLength);
@@ -76,25 +76,25 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
             bytes = BitConverter.GetBytes(reg.GetValue(rs2));
             reg.SetValue(rd, val);
 
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.Mem[addr + 0] = bytes[0];
+            reg.Mem[addr + 1] = bytes[1];
+            reg.Mem[addr + 2] = bytes[2];
+            reg.Mem[addr + 3] = bytes[3];
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -115,25 +115,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
-            Int32 val = BitConverter.ToInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val + (Int32)reg.GetValue(rs2));
-            reg.SetValue(rd, (UInt32)val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            UInt32 val = BitConverter.ToUInt32(bytes, 0);
+            reg.SetValue(rd, (UInt32)(val + reg.GetValue(rs2)));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -155,25 +149,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val ^ reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.SetValue(rd, val ^ reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -195,25 +183,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val & reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.SetValue(rd, val & reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -235,25 +217,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val | reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.SetValue(rd, val | reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -275,25 +251,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
-            UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes((Int32)val < (Int32)reg.GetValue(rs2) ? val : reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            Int32 val = BitConverter.ToInt32(bytes, 0);
+            reg.SetValue(rd, val < (Int32)reg.GetValue(rs2) ? (UInt32)val : reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -315,25 +285,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
-            UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes((Int32)val > (Int32)reg.GetValue(rs2) ? val : reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            Int32 val = BitConverter.ToInt32(bytes, 0);
+            reg.SetValue(rd, val > (Int32)reg.GetValue(rs2) ? (UInt32)val : reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -355,25 +319,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val < reg.GetValue(rs2) ? val : reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.SetValue(rd, val < reg.GetValue(rs2) ? val : reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
@@ -395,25 +353,19 @@ namespace RiscVCpu.LoadStoreUnit {
             }
 
             if (acquire) {
-                mem.Acquire(addr);
+                reg.Mem.Acquire(addr, 4);
             }
             byte[] bytes = new byte[4];
-            bytes[0] = mem[addr + 0];
-            bytes[1] = mem[addr + 1];
-            bytes[2] = mem[addr + 2];
-            bytes[3] = mem[addr + 3];
+            bytes[0] = reg.Mem[addr + 0];
+            bytes[1] = reg.Mem[addr + 1];
+            bytes[2] = reg.Mem[addr + 2];
+            bytes[3] = reg.Mem[addr + 3];
 
             UInt32 val = BitConverter.ToUInt32(bytes, 0);
-            bytes = BitConverter.GetBytes(val > reg.GetValue(rs2) ? val : reg.GetValue(rs2));
-            reg.SetValue(rd, val);
-
-            mem[addr + 0] = bytes[0];
-            mem[addr + 1] = bytes[1];
-            mem[addr + 2] = bytes[2];
-            mem[addr + 3] = bytes[3];
+            reg.SetValue(rd, val > reg.GetValue(rs2) ? val : reg.GetValue(rs2));
 
             if (release) {
-                mem.Release(addr);
+                reg.Mem.Release(addr);
             }
             reg.IncrementPc(insLength);
             return true;
